Add support for hashing files of the same size
Previously if a file was "touched" but not actually modified it would
show up as modified.  Now the behavior better mimics git and a file's
hash will be compared to that of what git thinks it is and only then
will it say it's modified.

It was also noticed that previously the modification time was only
accounting for seconds.  This got by most of the time, but quickly
exposed that this was too coarse of a comparison and could result in
skipping over files that really have changed.  Now the nanoseconds of a
file are also taken into account

+++ src/direntry.rs
@@ -20,7 +20,8 @@ impl Default for ObjectType {
 
 #[derive(PartialEq, Eq, Debug, Default, Clone)]
 pub struct FileStat {
-    pub mtime: u32,
+    // modified time in nanoseconds since the unix epoch
+    pub mtime: u128,
     pub size: u32,
 }
